% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-utility.R
\name{averageReplicateMatrices}
\alias{averageReplicateMatrices}
\title{Average a list of Matrices (cell wise)}
\usage{
averageReplicateMatrices(list, na.rm = FALSE)
}
\arguments{
\item{list}{a list of matrices of the same dims}

\item{na.rm}{should NA be removed before mean.}
}
\value{
the average matrix
}
\description{
Average a list of Matrices (cell wise)
}
\examples{
a <- matrix(1:6, 3,2)
b <- matrix(1:6+3, 3,2)
averageReplicateMatrices(list(a,b), na.rm=TRUE)
a[1,1] <- NA
averageReplicateMatrices(list(a,b), na.rm=TRUE)

}
