% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{computeFreqs}
\alias{computeFreqs}
\title{Compute frequences from list}
\usage{
computeFreqs(elementsIntersections)
}
\arguments{
\item{elementsIntersections}{a names list}
}
\value{
a 'data.frame' with the frequences
}
\description{
For internal use only. Compute frequences
}
