% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-survivalDataset-functions.R
\name{computeSimPATHyParameters}
\alias{computeSimPATHyParameters}
\title{Compute SimPATHy Parameters}
\usage{
computeSimPATHyParameters(chain, nodes, deInChain = NULL,
  deInPath = NULL, augmentedMu = 2, strong = FALSE)
}
\arguments{
\item{chain}{a chain o disregulation}

\item{nodes}{all the other nodes}

\item{deInChain}{the differentially expressed of the chain}

\item{deInPath}{the differentially expressed in Pathway}

\item{augmentedMu}{the magnetude of the mean inclement increment for DEGs}

\item{strong}{should be the increment really marked}
}
\value{
a list with the parameters
  \item{path}{the path}
  \item{min}{min}
  \item{max}{max}
  \item{prob}{prob}
  \item{mu1}{mu1}
  \item{mu2}{mu2}
}
\description{
Compute SimPATHy Parameters
}
