% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{computeSupertestFrequencies}
\alias{computeSupertestFrequencies}
\title{Compute supertest frequences}
\usage{
computeSupertestFrequencies(elementsIntersections)
}
\arguments{
\item{elementsIntersections}{a summary(msets) object}
}
\value{
a 'data.frame' with the frequences
}
\description{
For internal use only. Compute supertest frequences
}
