% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-utility.R
\name{convertIds}
\alias{convertIds}
\title{Ids converter}
\usage{
convertIds(from, keytype = "ENTREZID", to = "SYMBOL",
  annDB = "org.Hs.eg.db", graphiteStyle = FALSE)
}
\arguments{
\item{from}{ids to be converted}

\item{keytype}{the type of the starting id}

\item{to}{the arriving type}

\item{annDB}{\code{annotationDbi} package}

\item{graphiteStyle}{logical whether to expect graphite style ids}
}
\value{
vector of translated ids
}
\description{
For internal use only. Convert ids from to ids to.
}
