% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-survivalDataset-functions.R
\name{createPath}
\alias{createPath}
\title{Create a path from pathway}
\usage{
createPath(graph, start, end, mode = "out")
}
\arguments{
\item{graph}{Pathway graphite}

\item{start}{startig node}

\item{end}{end node}

\item{mode}{the direction to follow, out mean from source to dest}
}
\description{
Create a path from pathway
}
