% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{createSignificanceMask}
\alias{createSignificanceMask}
\title{Create a significance matrix}
\usage{
createSignificanceMask(list, thr = 0.05)
}
\arguments{
\item{list}{a list, each element a list of pvalues}

\item{thr}{threshold to consider significant}
}
\value{
a 'data.frame'. Each line a feature, columns are covariates: TRUE/FALSE significant
}
\description{
For internal use only. Create a data.frame of TRUE/FALSE
}
\examples{
l <- list(a=c(0.03, 0.06, 1), b=c(0.6, 0.01, 0.5))
createSignificanceMask(l)

}
