% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-relations.R
\name{downloadPathwayRelationFromReactome}
\alias{downloadPathwayRelationFromReactome}
\title{Download Reactome Pathway Relations}
\usage{
downloadPathwayRelationFromReactome(url = NULL, speciesAbbr = "HSA")
}
\arguments{
\item{url}{the location of the file. Can be local. If NULL pick the package reactome file.}

\item{speciesAbbr}{species acronim}
}
\value{
a data frame
}
\description{
Download Reactome Pathway Relations
}
\examples{
\dontrun{
url = "https://reactome.org/download/current/ReactomePathwaysRelation.txt"
downloadPathwayRelationFromReactome(url, speciesAbbr = "HSA")
}
}
