% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceEstimateFunctions.R
\name{estimateCov}
\alias{estimateCov}
\title{Estimate Covariance}
\usage{
estimateCov(exp1, exp2, shrink)
}
\arguments{
\item{exp1}{first numeric matrix}

\item{exp2}{second numeric matrix}

\item{shrink}{logical wheter to shrink the matrix}
}
\value{
a list
  \item{s1}{covariance of exp1}
  \item{s2}{covariance of exp2}
  \item{s}{combined covariance of exp1 and exp2}
}
\description{
For internal use only. Estimate 2 covariance matrixes
}
