% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceEstimateFunctions.R
\name{estimateExprCov}
\alias{estimateExprCov}
\title{Estimate Single Covariance Matrix}
\usage{
estimateExprCov(expr, shrink)
}
\arguments{
\item{expr}{a numeric matrix}

\item{shrink}{logical wheter to shrink the matrix}
}
\value{
a covariance matrix
}
\description{
For internal use only. Estimate Covariance from one matrix
}
