% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-survivalDataset-functions.R
\name{filterDirected}
\alias{filterDirected}
\title{Filter direct edges}
\usage{
filterDirected(graph, reverse = FALSE)
}
\arguments{
\item{graph}{a Pathway graph from graphite}

\item{reverse}{invert selection}
}
\value{
a Pathway with selected edges
}
\description{
Filter all directed edges
}
