% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-survivalDataset-functions.R
\name{fromChain2edgeList}
\alias{fromChain2edgeList}
\title{Convert a chain in an edgeList}
\usage{
fromChain2edgeList(chain)
}
\arguments{
\item{chain}{vector of node names}
}
\value{
edgeList
}
\description{
Convert a chain in an edgeList
}
