\name{general-utility}
\alias{convertIds}
\title{General utility functions}
\description{
This is a container for function usually very usefull in clip dinasty packages.
}
\usage{
convertIds(from, keytype="ENTREZID", to="SYMBOL", annDB="org.Hs.eg.db")
}
\arguments{
  \item{from}{starting ids}
  \item{keytype}{type of starting id}
  \item{to}{diestination id type}
  \item{annDB}{annotation db}
}
\value{
Functions meant for internal usage only.
}
\author{
Paolo Martini
}
