% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-relations.R
\name{getPathFathers}
\alias{getPathFathers}
\title{Retrieves pathways relatives}
\usage{
getPathFathers(pathway, hierarchyGraph, ord = 3, plot = F)
}
\arguments{
\item{pathway}{a pathway id}

\item{hierarchyGraph}{a igraph with pathway hierarchy}

\item{ord}{how far you need to go backward}

\item{plot}{plot relatives. For checking purpose}
}
\value{
a character vector with the relatives
}
\description{
For internal use only. Retrieves relatives given a pathway id.
}
\details{
Pathway Hierarchy is needed as igraph object.
}
