\name{graph-processing-functions}
\alias{mergeGraphitePathways}

\title{
Handle graph and sibilings
}
\description{
Multiple function to hanldle graph and graph structures.
}
\usage{
mergeGraphitePathways(pathways, id = "path1", title = "path1",
species = "hsapiens", database = "custom",
timestamp = Sys.Date())
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pathways}{pathways}
  \item{id}{id for merged pathway}
  \item{title}{title for merged pathway}
  \item{species}{the species for merged pathway}
  \item{database}{database name}
  \item{timestamp}{the creation date}
}
\value{
For internal usage only.
}
\author{
Paolo Martini
}
