% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-processing-functions.R
\name{extractCliquesFromDag}
\alias{extractCliquesFromDag}
\alias{mmmoralize}
\title{Extract the maximal cliques}
\usage{
extractCliquesFromDag(dag, root = NULL)

mmmoralize(graph)
}
\arguments{
\item{dag}{a Directed Aciclic Graph}

\item{root}{a node to use as root}

\item{graph}{a graphNEL object}
}
\value{
list of nodes cliques
}
\description{
For internal use only. Extract the cliques.

For internal use only. Force Moralization
}
\examples{
  graph <- gRbase::dag(c("me","ve"),c("me","al"),c("me","me"),
    c("ve","al"),c("al","an"),
    c("al","st"),c("an","st"))
  extractCliquesFromDag(graph)

}
