% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-relations.R
\name{id2name}
\alias{id2name}
\title{Convert id to pathway name}
\usage{
id2name(idList, namedVect)
}
\arguments{
\item{idList}{a list of pathway id}

\item{namedVect}{a named vector}
}
\value{
a character vector with the names
}
\description{
For internal use only. Retrieves name from pathway id.
}
\details{
You must provide a namedVect to be used as translator.
}
