% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-utility.R
\name{makeDegreesEven}
\alias{makeDegreesEven}
\title{Forse random nodes degress to be even}
\usage{
makeDegreesEven(degrees)
}
\arguments{
\item{degrees}{the nodes degrees, a named vector}
}
\value{
even nodes degrees
}
\description{
Forse random nodes degress to be even
}
