% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePositiveDefinite.R
\name{makePositiveDefinite}
\alias{makePositiveDefinite}
\title{Make positive and definite covariance matrix}
\usage{
makePositiveDefinite(m1, m2 = NULL, m3 = NULL, threshold = 0.1)
}
\arguments{
\item{m1}{matrix 1}

\item{m2}{matrix 2}

\item{m3}{matrix 3}

\item{threshold}{threshold of difference}
}
\value{
list with
  \item{m1}{the matrix m1 positive and definite}
  \item{m2}{the matrix m2 positive and definite}
  \item{m3}{the matrix m3 positive and definite}
  \item{correction}{the magneturde of the correction}
  \item{value}{the value}
}
\description{
Make positive and definite covariance matrix
}
