% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-survivalDataset-functions.R
\name{makeTheDataset}
\alias{makeTheDataset}
\title{Simulate a dataset with survival annotation}
\usage{
makeTheDataset(pathway, pathBoundaries, deInChain, deInPath, w, ann,
  augmentedMu = 2, strong = FALSE)
}
\arguments{
\item{pathway}{a Pathway}

\item{pathBoundaries}{the start and end point of the disregulated chain}

\item{deInChain}{the DEG in the aberrant signal chain}

\item{deInPath}{the DEG in Pathway}

\item{w}{the index of the path to choose}

\item{ann}{survival annotations}

\item{augmentedMu}{magnetude of mean increment}

\item{strong}{should the increment be really marked}
}
\value{
list
  \item{exprs}{expression}
  \item{annotation}{the survival annotation}
  \item{graph}{a graphNEL}
  \item{chain}{the chain of selectP}
}
\description{
To simulate a dataset we need a real pathway.
}
