% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-survivalDataset-functions.R
\name{makeUniformDataset}
\alias{makeUniformDataset}
\title{Make a dataset with no association between chains and survival}
\usage{
makeUniformDataset(pathway, ann)
}
\arguments{
\item{pathway}{a Pathway}

\item{ann}{survival annotations}
}
\description{
Make a dataset with no association between chains and survival
}
