% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-utility.R
\name{mapFromDict}
\alias{mapFromDict}
\title{Map ids using a dict.}
\usage{
mapFromDict(keys, dict)
}
\arguments{
\item{keys}{a vector of cluster Ids}

\item{dict}{a list with keys to values}
}
\value{
a list with value to keys
}
\description{
For internal use only.
}
