% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-processing-functions.R
\name{mergeGraphitePathways}
\alias{mergeGraphitePathways}
\title{Merge graphite Pathways}
\usage{
mergeGraphitePathways(pathways, id = "path1", title = "path1",
  species = "hsapiens", database = "custom", timestamp = Sys.Date())
}
\arguments{
\item{pathways}{a list of pathways.}

\item{id}{string for newly generated pathway id}

\item{title}{string for newly generated pathway title}

\item{species}{the species associated to the pathway}

\item{database}{the name of the databse}

\item{timestamp}{a date}
}
\value{
a \code{Pathway} object
}
\description{
For internal use only. Merge a list of Pathways from Graphite
}
