% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{minOrNA}
\alias{minOrNA}
\title{Minimum or NA}
\usage{
minOrNA(x)
}
\arguments{
\item{x}{a numeric}
}
\value{
a numeric. The minimum or NA
}
\description{
For internal use only. Get back minimum or NA.
}
\examples{
  minOrNA(c(1,5,0.1,NA))
  minOrNA(c(NA,NA,NA))

}
