% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{plotFrequencies}
\alias{plotFrequencies}
\title{Plot frequences}
\usage{
plotFrequencies(frequencies, manualColors = NULL, minSize = 4,
  maxSize = 20, width = 20, relMagnificationOfLegend = 0.5,
  lineSize = 1)
}
\arguments{
\item{frequencies}{a data.frame created from 'computeFreqs'}

\item{manualColors}{optional vector of colors}

\item{minSize}{the minimal fontsize. Maximal frequencies will be added for each class}

\item{maxSize}{the maximal fontsize dimension, all values above are clipped}

\item{width}{the numerbe of character to wrap the labels}

\item{relMagnificationOfLegend}{the relative magnification of the text of the legend}

\item{lineSize}{the thickness of the lines}
}
\description{
For internal use only. Plot supertest frequences
}
\examples{
df <- data.frame(category=c("talk", "too","mutch", "dear"),
  frequencies=c(1,2,1,3),
  class=rep("Mut",4), stringsAsFactors = FALSE)
plotFrequencies(df)

}
