% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphNEL_pathways_utility.R
\name{plotPathInGraphNEL}
\alias{plotPathInGraphNEL}
\title{Plot graphNEL with some nodes boerde highlighted}
\usage{
plotPathInGraphNEL(graph, path = NULL, singles = NULL,
  path_color = "red", singles_color = "green")
}
\arguments{
\item{graph}{graphNEL}

\item{path}{sequence of genes in a path}

\item{singles}{vector of singlets genes}

\item{path_color}{border color of the genes in the path}

\item{singles_color}{border color of the singlets}
}
\description{
Plot graphNEL with some nodes boerde highlighted
}
