% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{plotSupertest}
\alias{plotSupertest}
\title{Plot supertest}
\usage{
plotSupertest(mset)
}
\arguments{
\item{mset}{a 'msets' object}
}
\description{
For internal use only. Plot 'supertest' analysis.
}
