% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-utility.R
\name{randomExpression}
\alias{randomExpression}
\title{Create fake expression based on variance}
\usage{
randomExpression(n, nfeatures = 10)
}
\arguments{
\item{n}{the number of element (a.k.a. genes)}

\item{nfeatures}{the number of features (a.k.a. samples)}
}
\value{
dummy expression matrix
}
\description{
Create fake expression based on variance
}
\examples{
randomExpression(4)
}
