% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-relations.R
\docType{data}
\name{ReactomePathwaysRelation}
\alias{ReactomePathwaysRelation}
\title{This is a copy of the file
"https://reactome.org/download/current/ReactomePathwaysRelation.txt"
downloaded Dec 2018.}
\description{
A brand new file can be downloaded using:
 downloadPathwayRelationFromReactome
}
\keyword{data}
