% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-processing-functions.R
\name{removeSelfLoops}
\alias{removeSelfLoops}
\title{Remove self loops from a graphNEL}
\usage{
removeSelfLoops(graph)
}
\arguments{
\item{graph}{a graphNEL object}
}
\value{
a graphNEL object

#' @rdname  graph-processing
}
\description{
Remove the self loops that a present in the graph graphNEL object
}
\examples{
  graph <- gRbase::dag(c("me","ve"),c("me","al"),c("me","me"),
    c("ve","al"),c("al","an"),
    c("al","st"),c("an","st"))
  removeSelfLoops(graph)

}
