% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-utility.R
\name{reverseDict}
\alias{reverseDict}
\title{Invert The Dictionary}
\usage{
reverseDict(dict)
}
\arguments{
\item{dict}{a list with keys to values}
}
\value{
a list with value to keys
}
\description{
For internal use only. From key to value, to value to key.
}
