% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{runSuperTest}
\alias{runSuperTest}
\alias{createSignificanceMask}
\title{run supertest}
\usage{
runSuperTest(list, universeSize, thr = 0.05)

createSignificanceMask(list, thr = 0.05)
}
\arguments{
\item{list}{a list, each element a list of pvalues}

\item{universeSize}{the size of all the pvalues computed}

\item{thr}{threshold to consider significant}
}
\value{
a 'msets' object

a 'data.frame'. Each line a feature, columns are covariates: TRUE/FALSE significant
}
\description{
For internal use only. Perform 'supertest' analysis.

For internal use only. Create a data.frame of TRUE/FALSE
}
\examples{
l <- list(a=c(0.03, 0.06, 1), b=c(0.6, 0.01, 0.5))
createSignificanceMask(l)

}
