% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation-cnv_dataset_generation.R
\name{makeTheMutationDataset}
\alias{makeTheMutationDataset}
\alias{makeUniformMutationsDataset}
\title{Simulate a mutation dataset with survival annotation}
\usage{
makeTheMutationDataset(pathway, impacted_genes, patients_fractions, ann,
  omicName = "x", mutation_rate = 0.001)

makeUniformMutationsDataset(pathway, ann, omicName = "x",
  mutation_rate = 0.001)
}
\arguments{
\item{pathway}{a Pathway}

\item{impacted_genes}{the genes of a module with the aberrant signal chain}

\item{patients_fractions}{the fraction of patients with n mutated genes}

\item{ann}{survival annotations}

\item{omicName}{the name of the omics to consider}

\item{mutation_rate}{the basel mutation rate}
}
\value{
list
  \item{exprs}{expression}
  \item{annotation}{the survival annotation}
  \item{graph}{a graphNEL}
  \item{chain}{the chain of selectP}
}
\description{
To simulate a dataset we need a real pathway.

To simulate a dataset we need a real pathway.
}
