% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-survival-annotation.R
\name{simulateData}
\alias{simulateData}
\title{Simulate Survival Data}
\usage{
simulateData(beta, x, patientNum = 300, fUpTime = 1000, seed = NULL,
  row.names.pref = "P_")
}
\arguments{
\item{beta}{the beta values for covariates (see siple.surv.sim)}

\item{x}{the x values for covariates (see siple.surv.sim)}

\item{patientNum}{the number of patients to simulate}

\item{fUpTime}{the length of the followup}

\item{seed}{random seed setting}

\item{row.names.pref}{a prefix to row number to simulate real samples names}
}
\description{
Create survival censored data annotations
}
