% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-processing-functions.R
\name{subPathway}
\alias{subPathway}
\title{Subset of a graphite Pathway}
\usage{
subPathway(ids, pathway, and = TRUE)
}
\arguments{
\item{ids}{nodes ids}

\item{pathway}{a graphite Pathway object}

\item{and}{logical keep the edges between ids}
}
\value{
a \code{Pathway} object
}
\description{
For internal use only. Extract sub network from a Graphite Pathway.
}
