% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{summarizeOmicsResByMinPvalue}
\alias{summarizeOmicsResByMinPvalue}
\title{Summarize Omics Covaraites By Min Pvalue}
\usage{
summarizeOmicsResByMinPvalue(col, mat)
}
\arguments{
\item{col}{columns to extract from the line}

\item{mat}{the matrix to be summarized (were to extract lines and 'col')}
}
\value{
a summarized version of the matrix.
}
\description{
For internal use only. for each line extrac 'col' and get the minimum.
}
\examples{
  summarizeOmicsResByMinPvalue(2:3, mat=matrix(c(1,2,4,1,2,5), nrow=2))

}
