% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-functions-with-resampling.R,
%   R/pca-functions.R
\name{topoCompPCsWithResampling}
\alias{topoCompPCsWithResampling}
\alias{sparseCompPCsWithResampling}
\alias{compPCsWithResampling}
\alias{computePCs}
\alias{topoCompPCs}
\alias{sparseCompPCs}
\alias{compPCs}
\title{Topological PCA}
\usage{
topoCompPCsWithResampling(resampling, exp, shrink, cliques, k)

sparseCompPCsWithResampling(resampling, exp, shrink, k)

compPCsWithResampling(resampling, exp, shrink, k)

computePCs(exp, shrink = FALSE, method = c("regular", "topological",
  "sparse"), cliques = NULL, maxPCs = 3)

topoCompPCs(exp, shrink, cliques, k)

sparseCompPCs(exp, shrink, k)

compPCs(exp, shrink, k)
}
\arguments{
\item{resampling}{list of resampled columns}

\item{exp}{a matrix}

\item{shrink}{logical, whether to shrink or not.}

\item{cliques}{the pathway topology summarized in a list of cliques}

\item{k}{the number of components to use}

\item{method}{one of 'regular', 'topological' and 'sparse'}

\item{maxPCs}{the maximum number of PCs to consider}
}
\value{
a list with the following elements:
  \item{x}{the computed PCs}
  \item{sdev}{the standard deviation captured by the PCs}
  \item{loadings}{the loadings}

a list with the following elements:
  \item{x}{the computed PCs}
  \item{sdev}{the standard deviation captured by the PCs}
  \item{loadings}{the loadings}

a list with the following elements:
  \item{x}{the computed PCs}
  \item{sdev}{the standard deviation captured by the PCs}
  \item{loadings}{the loadings}

a list with the following elements:
  \item{x}{the computed PCs}
  \item{sdev}{the standard deviation captured by the PCs}
  \item{loadings}{the loadings}
}
\description{
Topological PCA

Sparse PCA

Regular PCA

For internal use only. Performs Principal Componenent analysis.

Topological PCA

Sparse PCA

Regular PCA
}
\details{
Three methods are implemented:
  * regular: a regular PCA ('prcomp')
  * topological: PCA using a pathway topology.
  * sparse: sparse PCA analysis implemented by 'elasticnet'
}
\examples{
  fakeExp <- randomExpression(4)
  computePCs(t(fakeExp))

}
