% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{computeSupertestFrequencies}
\alias{computeSupertestFrequencies}
\alias{computeFreqs}
\title{Compute supertest frequences}
\usage{
computeSupertestFrequencies(elementsIntersections, sep = ".")

computeFreqs(elementsIntersections)
}
\arguments{
\item{elementsIntersections}{a summary(msets) object}

\item{sep}{the field separator of the class}
}
\value{
a 'data.frame' with the frequences
}
\description{
For internal use only. Compute supertest frequences

For internal use only. Compute  frequences
}
