% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-relations.R
\name{downloadPathwayRelationFromReactome}
\alias{downloadPathwayRelationFromReactome}
\title{Download Reactome Pathway Relations}
\usage{

  downloadPathwayRelationFromReactome(url = "https://reactome.org/download/current/ReactomePathwaysRelation.txt",
  speciesAbbr = "HSA")
}
\arguments{
\item{url}{the location of the file. Can be local.}

\item{speciesAbbr}{species acronim}
}
\value{
a data frame
}
\description{
Download Reactome Pathway Relations
}
