% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphite_pathways_utilities.R
\name{filterDirected}
\alias{filterDirected}
\alias{createPath}
\alias{fromChain2edgeList}
\title{Filter direct edges}
\usage{
filterDirected(graph, reverse = FALSE)

createPath(graph, start, end, mode = "out")

fromChain2edgeList(chain)
}
\arguments{
\item{graph}{a Pathway graph from graphite}

\item{reverse}{invert selection}

\item{start}{startig node}

\item{end}{end node}

\item{mode}{the direction to follow, out mean from source to dest}

\item{chain}{vector of node names}
}
\value{
a Pathway with selected edges

edgeList
}
\description{
Filter all directed edges
}
\examples{
  if (require(graphite)) {
    p <- pathways("hsapiens", "kegg")[["Cell cycle"]]
    createPath(p, "ENTREZID:25", "ENTREZID:7029")
  }

}
