% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-graphite-pathways.R
\name{mapPathwaysIDfromGraphite}
\alias{mapPathwaysIDfromGraphite}
\title{Map Pathways ID from Graphite}
\usage{
mapPathwaysIDfromGraphite(pathways, pathwayNames = NULL)
}
\arguments{
\item{pathways}{a PathwayList object}

\item{pathwayNames}{in not NULL, a subset of pathway to extract}
}
\value{
a data frame, id and pathway name
}
\description{
For internal use only. Retrieve pathway id and names from Pathways object.
}
\examples{
  if (require(graphite)){
    pathways <- pathways("hsapiens", "kegg")
    mapPathwaysIDfromGraphite(pathways, names(pathways)[1:10])
  }
}
