% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{plotSupertestFrequencies}
\alias{plotSupertestFrequencies}
\title{Plot supertest frequences}
\usage{
plotSupertestFrequencies(supertestFreq, manualColors = NULL)
}
\arguments{
\item{supertestFreq}{a data.frame created from 'computeSupertestFrequencies'}

\item{manualColors}{optional vector of colors}
}
\description{
For internal use only. Plot supertest frequences
}
