% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superExactTest.R
\name{plotSupertestFrequencies}
\alias{plotSupertestFrequencies}
\title{Plot supertest frequences}
\usage{
plotSupertestFrequencies(supertestFreq)
}
\arguments{
\item{supertestFreq}{a data.frame created from 'computeSupertestFrequencies'}
}
\description{
For internal use only. Plot supertest frequences
}
