% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-utility.R
\name{randomExpression}
\alias{randomExpression}
\title{Create fake expression}
\usage{
randomExpression(component_x = 2, component_y = 2, dimension = c(10, 20))
}
\arguments{
\item{component_x}{the number of component in x}

\item{component_y}{the number of component in y}

\item{dimension}{the final dimension of the expression matrix}
}
\value{
dummy expression matrix
}
\description{
Create fake expression
}
\examples{
randomExpression()
}
