% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-survivalDataset-functions.R
\name{simulateData}
\alias{simulateData}
\title{Simulate Survival Data}
\usage{
simulateData(beta, x, patientNum = 300, fUpTime = 1000, seed = NULL)
}
\arguments{
\item{beta}{the beta values for covariates (see siple.surv.sim)}

\item{x}{the x values for covariates (see siple.surv.sim)}

\item{patientNum}{the number of patients to simulate}

\item{fUpTime}{the length of the followup}

\item{seed}{random seed setting}
}
\description{
Create survival censored data annotations
}
